+++ services/src/main/java/org/keycloak/services/managers/AuthenticationManager.java
@@ -1198,8 +1198,7 @@ public class AuthenticationManager {
                 continue;
             }
 
-            // we need to add dynamic scopes with params to the scopes to consent every time for now
-            if (grantedConsent == null || !grantedConsent.isClientScopeGranted(clientScope) || isDynamicScopeWithParam(authDetails)) {
+            if (grantedConsent == null || !grantedConsent.isClientScopeGranted(clientScope) ) {
                 clientScopesToDisplay.add(authDetails);
             }
         }