+++ testsuite/integration-arquillian/tests/base/src/test/java/org/keycloak/testsuite/oauth/OAuthGrantTest.java
@@ -38,6 +38,7 @@ import org.keycloak.representations.idm.EventRepresentation;
 import org.keycloak.representations.idm.ProtocolMapperRepresentation;
 import org.keycloak.representations.idm.RealmRepresentation;
 import org.keycloak.representations.idm.RoleRepresentation;
+import org.keycloak.representations.idm.UserRepresentation;
 import org.keycloak.testsuite.AbstractKeycloakTest;
 import org.keycloak.testsuite.AssertEvents;
 import org.keycloak.testsuite.admin.ApiUtil;
@@ -52,15 +53,21 @@ import org.keycloak.testsuite.util.OAuthClient;
 import org.keycloak.testsuite.util.ProtocolMapperUtil;
 import org.openqa.selenium.By;
 
+import java.util.AbstractMap;
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 import javax.ws.rs.core.Response;
 
 import static org.junit.Assert.assertEquals;
 import static org.keycloak.testsuite.admin.AbstractAdminTest.loadJson;
 import static org.keycloak.testsuite.admin.ApiUtil.findClientByClientId;
+import static org.keycloak.testsuite.util.ProtocolMapperUtil.createClaimMapper;
+import static org.keycloak.testsuite.util.ProtocolMapperUtil.createHardcodedClaim;
 
 /**
  * @author <a href="mailto:vrockai@redhat.com">Viliam Rockai</a>
@@ -336,6 +343,11 @@ public class OAuthGrantTest extends AbstractKeycloakTest {
             put(ClientScopeModel.IS_DYNAMIC_SCOPE, "true");
             put(ClientScopeModel.DYNAMIC_SCOPE_REGEXP, "foo-dynamic-scope:*");
         }});
+
+        List<ProtocolMapperRepresentation> mappers = new ArrayList<>();
+        mappers.add(createClaimMapper("foo-dynamic-scope-mapper", "foo-dynamic", "foo-dynamic-scope", "String", true, true,true));
+        scope.setProtocolMappers(mappers);
+
         Response response = appRealm.clientScopes().create(scope);
         String dynamicFooScopeId = ApiUtil.getCreatedId(response);
         response.close();
@@ -343,6 +355,11 @@ public class OAuthGrantTest extends AbstractKeycloakTest {
 
         // Add clientScope as optional to client
         thirdParty.addOptionalClientScope(dynamicFooScopeId);
+
+        UserRepresentation userRep = appRealm.users().search("test-user@localhost").get(0);
+        List<String> attributeValues = Stream.of("withparam","param").collect(Collectors.toList());
+        userRep.setAttributes(Stream.of(new AbstractMap.SimpleEntry<>("foo-dynamic",attributeValues)).collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue)));
+        appRealm.users().get(userRep.getId()).update(userRep);
 
         // Assert clientScope not on grant screen when not requested
         oauth.clientId(THIRD_PARTY_APP);
@@ -360,6 +377,13 @@ public class OAuthGrantTest extends AbstractKeycloakTest {
 
         String code = new OAuthClient.AuthorizationEndpointResponse(oauth).getCode();
         OAuthClient.AccessTokenResponse res = oauth.doAccessTokenRequest(code, "password");
+        Assert.assertNotNull(res.getAccessToken());
+        AccessToken token = oauth.verifyToken(res.getAccessToken());
+        Assert.assertNotNull(token);
+        //foo-dynamic-scope must contain only withparam
+        List<String> claimValue =(List) token.getOtherClaims().get("foo-dynamic-scope");
+        assertEquals(claimValue.size(),1);
+        assertEquals(claimValue.get(0),"withparam");
 
         events.expectCodeToToken(loginEvent.getDetails().get(Details.CODE_ID), loginEvent.getSessionId())
                 .client(THIRD_PARTY_APP)
@@ -375,15 +399,9 @@ public class OAuthGrantTest extends AbstractKeycloakTest {
         // login again to check whether the Dynamic scope and only the dynamic scope is requested again
         oauth.scope("foo-dynamic-scope:withparam");
         oauth.doLogin("test-user@localhost", "password");
-        grantPage.assertCurrent();
-        grants = grantPage.getDisplayedGrants();
-        Assert.assertEquals(1, grants.size());
-        Assert.assertTrue(grants.contains("foo-dynamic-scope: withparam"));
-        grantPage.accept();
-
         events.expectLogin()
                 .client(THIRD_PARTY_APP)
-                .detail(Details.CONSENT, Details.CONSENT_VALUE_CONSENT_GRANTED)
+                .detail(Details.CONSENT, Details.CONSENT_VALUE_PERSISTED_CONSENT)
                 .assertEvent();
 
         // Revoke
@@ -394,6 +412,8 @@ public class OAuthGrantTest extends AbstractKeycloakTest {
 
         // cleanup
         oauth.scope(null);
+        userRep.getAttributes().remove("foo-dynamic");
+        appRealm.users().get(userRep.getId()).update(userRep);
         thirdParty.removeOptionalClientScope(dynamicFooScopeId);
     }