+++ testsuite/integration-arquillian/tests/base/src/test/java/org/keycloak/testsuite/oauth/ResourceOwnerPasswordCredentialsGrantTest.java
@@ -50,6 +50,7 @@ import org.keycloak.representations.RefreshToken;
 import org.keycloak.representations.idm.ClientRepresentation;
 import org.keycloak.representations.idm.ClientScopeRepresentation;
 import org.keycloak.representations.idm.EventRepresentation;
+import org.keycloak.representations.idm.ProtocolMapperRepresentation;
 import org.keycloak.representations.idm.RealmRepresentation;
 import org.keycloak.representations.idm.UserRepresentation;
 import org.keycloak.testsuite.AbstractKeycloakTest;
@@ -68,14 +69,21 @@ import org.keycloak.testsuite.util.UserManager;
 
 import java.io.UnsupportedEncodingException;
 import java.security.Security;
+import java.util.AbstractMap;
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.LinkedList;
 import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
+import static org.keycloak.testsuite.util.ProtocolMapperUtil.createClaimMapper;
+import static org.keycloak.testsuite.util.ProtocolMapperUtil.createHardcodedClaim;
 
 import javax.validation.constraints.AssertTrue;
 import javax.ws.rs.core.HttpHeaders;
@@ -199,7 +207,17 @@ public class ResourceOwnerPasswordCredentialsGrantTest extends AbstractKeycloakT
             put(ClientScopeModel.DYNAMIC_SCOPE_REGEXP, "dynamic-scope:*");
         }});
         clientScope.setProtocol(OIDCLoginProtocol.LOGIN_PROTOCOL);
+        // create the attribute mapper
+        List<ProtocolMapperRepresentation> mappers = new ArrayList<>();
+        mappers.add(createClaimMapper("dynamic-scope-mapper", "dynamic", "dynamic-scope", "String", true, true,true));
+        clientScope.setProtocolMappers(mappers);
         RealmResource realmResource = adminClient.realm("test");
+
+        UserRepresentation userRep = realmResource.users().search("direct-login").get(0);
+        List<String> attributeValues = Stream.of("123","456").collect(Collectors.toList());
+        userRep.setAttributes(Stream.of(new AbstractMap.SimpleEntry<>("dynamic",attributeValues)).collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue)));
+        realmResource.users().get(userRep.getId()).update(userRep);
+
         try(Response response = realmResource.clientScopes().create(clientScope)) {
             String scopeId = ApiUtil.getCreatedId(response);
             getCleanup().addClientScopeId(scopeId);
@@ -209,12 +227,12 @@ public class ResourceOwnerPasswordCredentialsGrantTest extends AbstractKeycloakT
             resourceOwnerPublicClient.addOptionalClientScope(scopeId);
         }
 
-        oauth.scope("dynamic-scope:123");
+        oauth.scope("dynamic-scope:123 dynamic-scope:789");
         oauth.clientId("resource-owner-public");
 
         OAuthClient.AccessTokenResponse response = oauth.doGrantAccessTokenRequest("secret", "direct-login", "password");
 
-        assertTrue(response.getScope().contains("dynamic-scope:123"));
+        assertTrue(response.getScope().contains("dynamic-scope:123 dynamic-scope:789"));
 
         assertEquals(200, response.getStatusCode());
 
@@ -234,12 +252,18 @@ public class ResourceOwnerPasswordCredentialsGrantTest extends AbstractKeycloakT
                 .removeDetail(Details.CONSENT)
                 .assertEvent();
 
-        assertTrue(accessToken.getScope().contains("dynamic-scope:123"));
+        assertTrue(accessToken.getScope().contains("dynamic-scope:123 dynamic-scope:789"));
+        List<String> claimValue =(List) accessToken.getOtherClaims().get("dynamic-scope");
+        assertEquals(claimValue.size(),1);
+        assertEquals(claimValue.get(0),"123");
+
+        userRep.getAttributes().remove("dynamic");
+        adminClient.realm("test").users().get(userRep.getId()).update(userRep);
     }
 
     @Test
     @EnableFeature(value = Profile.Feature.DYNAMIC_SCOPES, skipRestart = true)
-    public void grantAccessTokenWithUnassignedDynamicScope() throws Exception {;
+    public void grantAccessTokenWithUnassignedDynamicScope() throws Exception {
         oauth.scope("unknown-scope:123");
         oauth.clientId("resource-owner-public");
         OAuthClient.AccessTokenResponse response = oauth.doGrantAccessTokenRequest("secret", "direct-login", "password");